refactor: extract @Syncable macro-support functions out of Core.swift
Core.swift had become a 1300-line grab-bag. As a first step, move the
runtime SPI the @Syncable macro generates calls to — the
syncApplyTo{One,Many}{ForeignKey,NestedObject} family, exportEncodeValue,
exportSetValue, and their exclusive private helpers — into
SyncableMacroSupport.swift.

resolveIdentity / syncIdentityKey stay in Core (also used by
SyncRelationshipLookupCache) and widen private -> internal.

Pure relocation, no behavior change: Core 1300 -> 723 lines. Build +
full suite green. (Homing these as non-free functions is a follow-up.)
